% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_regiones.R
\name{code_regiones}
\alias{code_regiones}
\title{This function matches municipalities by name to their corresponding region}
\usage{
code_regiones(dfName, varName)
}
\arguments{
\item{dfName}{A dataframe containing a variable with municipalities}

\item{varName}{A variable containing municipalities, in quotation marks.}
}
\value{
A character vector.
}
\description{
This function matches municipalities by name to their corresponding region
}
\examples{
df <- data.frame(comunas=c("ñuñoa", "Peñalolén", "PEñaflor", "Las condes"))
clean_df <- clean_comunas(df, "comunas")
clean_df_regiones <- code_regiones(clean_df, "comunas")
}
