% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_comunas.R
\name{code_comunas}
\alias{code_comunas}
\title{This function matches municipalities by name to their corresponding INE code}
\usage{
code_comunas(dfName, varName)
}
\arguments{
\item{dfName}{A dataframe containing a variable with municipalities}

\item{varName}{A variable containing municipalities, in quotation marks.}
}
\value{
A character vector.
}
\description{
This function matches municipalities by name to their corresponding INE code
}
\examples{
df <- data.frame(comunas=c("ñuñoa", "Peñalolén", "PEñaflor", "Las condes"))
clean_df <- clean_comunas(df, "comunas")
code_comunas(clean_df, "comunas")
}
